Add basic game engine
