+++ src/engine/view.ts
@@ -1,0 +1,5 @@
+export abstract class View {
+    public abstract init(): Promise<void>;
+    public abstract end(): void
+    public abstract update(dt: number): void
+}
